Squashed commit of the following:
commit a368b005a440121fb716a364c4682f3bef4dbd9d
Author: robbyxp1 <robxp@hotmail.com>
Date:   Wed May 15 16:07:11 2024 +0100

    Rework FSD spec to get data from itemdata, update itemdata again

commit 648d4895f82eef18657e5ff7ebaeea0270eed6a2
Author: robbyxp1 <robxp@hotmail.com>
Date:   Wed May 15 15:10:59 2024 +0100

    Full constructor/print out

commit 2d7c101c066b1e541421b5701800273fdc395775
Author: robbyxp1 <robxp@hotmail.com>
Date:   Tue May 14 15:28:05 2024 +0100

    Move const

commit 9d7544fbe29b303a09f74772ba8dec9d71efed22
Author: robbyxp1 <robxp@hotmail.com>
Date:   Tue May 14 15:06:39 2024 +0100

    Constructors part1

commit f0b8f12c830d9e976b7163dd520507d5895e9b76
Author: robbyxp1 <robxp@hotmail.com>
Date:   Tue May 14 14:43:55 2024 +0100

    Normalised vs EDSY

commit ba97f869864ad50e718dd93e5535d634e2ea6124
Author: robbyxp1 <robxp@hotmail.com>
Date:   Tue May 14 14:39:09 2024 +0100

    Renable some lines

commit a87313eabe9d3fd6c9bdac4cdbd8968c667171d4
Author: robbyxp1 <robxp@hotmail.com>
Date:   Tue May 14 14:07:16 2024 +0100

    Push half update

+++ EliteDangerous/EliteDangerous/EngineeringData.cs
@@ -200,7 +200,7 @@ namespace EliteDangerousCore
                 newmodule.Mass = mod.Value;
             mod = FindModification("FSDOptimalMass");
             if (mod != null)
-                newmodule.OptMass = mod.Value;
+                newmodule.OptMass = (int)mod.Value;
  
             return newmodule;
         }
